+++ src/gemseo_umdo/formulations/sequential_sampling.py
@@ -36,6 +36,8 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 from typing import Any
 
+from gemseo import SEED
+from gemseo.algos.doe.lib_openturns import OpenTURNS
 from gemseo.core.discipline import MDODiscipline
 
 from gemseo_umdo.formulations.sampling import Sampling
@@ -73,9 +75,9 @@ class SequentialSampling(Sampling):
         objective_statistic_parameters: Mapping[str, Any] | None = None,
         maximize_objective: bool = False,
         grammar_type: MDODiscipline.GrammarType = MDODiscipline.GrammarType.JSON,
-        algo: str = "OT_OPT_LHS",
+        algo: str = OpenTURNS.OT_LHSO,
         algo_options: Mapping[str, Any] | None = None,
-        seed: int = 1,
+        seed: int = SEED,
         **options: Any,
     ) -> None:
         """